// models/kolaborasi/project/ListProjectModel.ts
export class ListProjectModel {

    projectId?: string;
    projectTitle?: string;
    projectLogo?: string;
    creatorName?: string;
    colorId?: string;
    color?: string;
    totalMessage?: number;
    isMessageUnread?: number;
    messageUnreadTotal?: number;
    taskTotal?: number;
    taskNew?: number;
    taskProgress?: number;
    taskPostponed?: number;
    taskDone?: number;
    taskInit?: number;
    colaboratorTotal?: number;
    projectCreated?: string
}